+++ Source/GameObjects/EnemyBullet.cpp
@@ -1,5 +1,6 @@
 #include "../pch.h"
 #include "EnemyBullet.h"
+#include "../globals.h"
 
 EnemyBullet::EnemyBullet() :Bullet(SimpleMath::Vector2(0, 0)) {
 }
@@ -12,10 +13,10 @@ EnemyBullet::EnemyBullet(Vector2 & position) : Bullet(position) {
 EnemyBullet::~EnemyBullet() {
 }
 
-#include "../globals.h"
+
 void EnemyBullet::update(double deltaTime) {
 
-	Vector2 newpos = position + direction*deltaTime*bulletSpeed;
+	Vector2 newpos = position + direction*FLOAT(deltaTime*bulletSpeed);
 	setPosition(newpos);
 	if (position.x < 0 || position.x > Globals::WINDOW_WIDTH
 		|| position.y < 0 || position.y > Globals::WINDOW_HEIGHT)
